@@ -1,8 +1,10 @@
 {
   "name": "testausserveri.fi",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
+    "prebuild": "npm run dump-licenses",
     "build": "next build",
     "postbuild": "next-sitemap",
     "start": "next start",
@@ -12,34 +14,46 @@
     "dump-licenses": "npx license-checker --json --production --customPath dependency-license-format.json > dependency-licenses.json && sed -i -e 's/\"Custom: .*\"/\"\"/g' dependency-licenses.json"
   },
   "dependencies": {
+    "@plaiceholder/next": "^3.0.0",
     "@skyra/discord-components-core": "^3.6.0",
     "@skyra/discord-components-react": "^3.6.0",
     "@splinetool/react-spline": "^2.2.6",
     "@splinetool/runtime": "^0.9.268",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
     "hamburger-react": "^2.5.0",
-    "next": "^13.2.4",
-    "next-mdx-remote": "^4.4.1",
+    "next": "^14.2.5",
+    "next-mdx-remote": "^5.0.0",
     "next-sitemap": "^4.1.8",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "plaiceholder": "^3.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^4.8.0",
-    "react-tooltip": "^5.10.1-beta.4",
-    "styled-components": "^5.3.9"
+    "react-photo-album": "^3.0.2",
+    "react-tooltip": "^5.27.1",
+    "remark-gfm": "^4.0.0",
+    "rss-parser": "^3.13.0",
+    "sass": "^1.77.8",
+    "sharp": "^0.32.6",
+    "styled-components": "^5.3.9",
+    "yet-another-react-lightbox": "^3.21.4"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "^7.0.4",
-    "@storybook/addons": "^7.0.4",
-    "@storybook/nextjs": "^7.0.4",
-    "@storybook/react": "^7.0.4",
-    "@storybook/theming": "^7.0.4",
+    "@chromatic-com/storybook": "^1.6.1",
+    "@storybook/addon-essentials": "^8.2.5",
+    "@storybook/addons": "^7.6.17",
+    "@storybook/nextjs": "^8.2.5",
+    "@storybook/react": "^8.2.5",
+    "@storybook/theming": "^8.2.5",
     "@types/node": "^18.15.3",
     "@types/react": "^18.0.28",
     "@types/styled-components": "^5.1.26",
-    "eslint": "8.36.0",
-    "eslint-config-next": "13.2.4",
-    "eslint-plugin-storybook": "^0.6.11",
-    "storybook": "^7.0.4",
+    "eslint-config-next": "^14.2.5",
+    "eslint-plugin-storybook": "^0.8.0",
+    "file-loader": "^6.2.0",
+    "license-checker": "^25.0.1",
+    "storybook": "^8.2.5",
     "storybook-css-modules-preset": "^1.1.1",
-    "typescript": "^5.0.2"
+    "typescript": "^5.5.3"
   }
 }
