@@ -1,30 +1,19 @@
-const { PHASE_DEVELOPMENT_SERVER } = require('next/constants')
+// @ts-check
+import withPlaiceholder from "@plaiceholder/next";
 
-// The lines API_SERVER_DEV and URL_DEV are the most significant to you
-const API_SERVER_DEV = `https://api.testausserveri.fi`
-const API_SERVER_PROD = `https://api.testausserveri.fi`
-
-const URL_DEV = `http://localhost:3000`
-const URL_PROD = `https://testausserveri.fi`
-
-const MEDIA_API_SERVER = 'https://api.testausserveri.fi'
-/// ---
-
-module.exports = (phase) => ({
+/**
+ * @type {import('next').NextConfig}
+ */
+const config = {
   reactStrictMode: true,
   images: {
-    domains: ['localhost', 'api.testausserveri.fi', 'avatars.githubusercontent.com'],
-  },
-  env: {
-    NEXT_PUBLIC_API_SERVER_MEDIA: MEDIA_API_SERVER,
-    NEXT_PUBLIC_API_SERVER: (phase === PHASE_DEVELOPMENT_SERVER ? API_SERVER_DEV : API_SERVER_PROD),
-    NEXT_PUBLIC_URL: (phase === PHASE_DEVELOPMENT_SERVER ? URL_DEV : URL_PROD)
+    domains: ['localhost', 'api.testausserveri.fi', 'avatars.githubusercontent.com', 'testausauto.fi'],
   },
   async rewrites() {
 		return [
 			{
 				source: '/api/v1/:path*',
-				destination: (phase === PHASE_DEVELOPMENT_SERVER ? API_SERVER_DEV : API_SERVER_PROD) + '/v1/:path*',
+				destination: process.env.NEXT_PUBLIC_API_SERVER + '/v1/:path*',
 			},
 		]
 	},
@@ -47,5 +36,24 @@ module.exports = (phase) => ({
       { "source": "/host/privacy", "destination": "/privacy/host", permanent: true }
     ];
   },
-})
-
+  webpack: (config, options) => {
+    config.module.rules.push({ 
+      test: /\.mdx$/, use: 'raw-loader' 
+    })
+    const prefix = config.assetPrefix ?? config.basePath ?? '';
+    config.module.rules.push({
+      test: /\.mp4$/,
+      use: [{
+        loader: 'file-loader',
+        options: {
+          publicPath: `${prefix}/_next/static/media/`,
+          outputPath: `${config.dev ? '' : '../'}${config.isServer ? '../' : ''}static/media/`,
+          name: '[name].[hash:8].[ext]',
+        },
+      }],
+    });
+    return config
+  }
+};
+ 
+export default withPlaiceholder(config);
