@@ -1,8 +1,9 @@
+"use client";
+
 import styles from './Explanation.module.css'
 import { RiQuestionLine } from "react-icons/ri"
 import { PropsWithChildren, useId } from 'react';
 import { Tooltip } from 'react-tooltip';
-import 'react-tooltip/dist/react-tooltip.css'
 
 export type ExplanationProps = PropsWithChildren<{}>
 
@@ -11,7 +12,7 @@ export function Explanation({ children }: ExplanationProps) {
 
     return (
         <>
-            <Tooltip id={id} style={{maxWidth: "300px"}} className="tip">
+            <Tooltip id={id} className="tip">
                 {children}
             </Tooltip>
             <span className={styles.explanation} data-tooltip-id={id}><RiQuestionLine /></span>
