@@ -132,7 +132,7 @@ export default function HostPrivacy({ copyrightYear }: InferGetServerSidePropsTy
           <p>Rekisteröity voi tehdä henkilötietojen käsittelystä valituksen valvovalle viranomaiselle, joka on tietosuojavaltuutettu.</p>
         </Section>
       </Content>
-      <Footer copyrightYear={copyrightYear} />
+      <Footer />
     </div>
   )
 }
