@@ -21,7 +21,8 @@ import { getMemberAvatarUrl } from '../../utils/Member'
 import { DetailedProject, ShallowProject } from "../../utils/types";
 import Link from 'next/link'
 import Image from 'next/image'
-import { Key } from 'react'
+import { mdxComponents } from '../../components/mdx/mdxComponents'
+import { FadeBackground } from '../../components/FadeBackground/FadeBackground'
 
 const Layout = styled.div`
   margin-top: 2rem;
@@ -50,18 +51,6 @@ const Layout = styled.div`
     }
   } 
 `
-
-const AvatarRowExtended = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  span {
-    margin-left: .5rem;
-    font-size: 0.95em;
-    flex: 1;
-  }
-`
-
 const ProjectLinks = styled.ul`
   list-style-type: none;
   margin: 0;
@@ -86,36 +75,6 @@ const ProjectLinks = styled.ul`
     font-size: 1rem;
     font-family: 'Poppins';
   }
-`
-
-const FadeBackground = styled.div`
-  &::before {
-    content: '';
-    background-image: var(--bg);
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 35rem;
-    background-size: cover;
-    opacity: 0.3;
-    z-index: -10;
-    @-moz-document url-prefix() {
-      opacity: 0.15;
-    }
-  }
-  &::after {
-    content: ' ';
-    width: 100%;
-    bottom: 0;
-    position: absolute;
-    left: 0;
-    height: 35rem;
-    z-index: -1;
-    background: linear-gradient(180deg, rgba(13, 13, 13, 0) 0%, rgba(0,0,0,0.7) 40.67%, #0D0D0D 96.87%);
-    backdrop-filter: blur(5px);
-    top: 0px;
-  }  
 `
 
 const RepositoryReadme = styled.div`
@@ -168,74 +127,6 @@ const ProjectLinkTitleContainer = styled.div`
   align-items: center;
 `
 
-const Blockquote = styled.blockquote`
-  text-align: center;
-  margin: 3rem 0;
-  color: rgba(255,255,255, 0.7);
-`
-
-const MdxImageParent = styled.div`
-  &.inline {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    max-height: 500px;
-    @media only screen and (max-width: 800px) {
-      flex-direction: column;
-      max-height: 100%;
-      gap: 1rem;
-    }
-    .img-wrapper {
-      flex: 1;
-      height: 100%;
-      margin-right: 10px; 
-    }
-    img {
-      height: 100%;
-      width: auto;
-      object-fit: cover;
-    }
-  }
-  img {
-    max-height: 500px;
-    width: auto;
-    max-width: 100%;
-    margin: 0 auto;
-    display: block;
-    border-radius: 0.5rem;
-  }
-  small {
-    text-align: center;
-    width: 100%;
-    display: block;
-    margin: .5rem 0 1rem 0;
-  }
-`
-
-type MdxImageProps = {
-  src: string,
-  caption?: string
-}
-
-const MdxImage = (props: MdxImageProps) => {
-  if (Array.isArray(props.src)) {
-    return <MdxImageParent className="inline">
-      {props.src.map((url: string, index: number)=> (
-        <div key={url as Key} className="img-wrapper">
-          <img src={url} alt={`Kuva ${index} ${props?.caption ? ": " + props.caption : ""}`} />
-        </div>
-      ))}
-    </MdxImageParent>
-  } else {
-    return <MdxImageParent>
-      <img src={props.src} alt={props?.caption} />
-      {props.caption ?
-        <small>{props.caption}</small>
-      : null}
-    </MdxImageParent>
-  }
-}
-const components = { Blockquote, Image: MdxImage}
 
 export default function ProjectPage({ projectData: project, mdxSerialized, suggestedProjectsData: suggestedProjects, copyrightYear }: InferGetStaticPropsType<typeof getStaticProps>) {
   const cover = project.media.find(item => item.cover === true)
@@ -251,7 +142,7 @@ export default function ProjectPage({ projectData: project, mdxSerialized, sugge
         <meta name="twitter:card" content="summary_large_image"></meta>
       </Head>
       {/* @ts-ignore */}
-      <FadeBackground style={cover?.filename ? { "--bg": `url('${getProjectMediaUrl(cover.filename)}')` } : {}}>
+      <FadeBackground url={getProjectMediaUrl(cover.filename || "")}>
         <Content wider>
           <Breadcrumbs
             route={[
@@ -273,7 +164,7 @@ export default function ProjectPage({ projectData: project, mdxSerialized, sugge
               </P>
 
               {project.description.full ?
-                <MDXRemote {...mdxSerialized.fullDescription} components={components}/>
+                <MDXRemote {...mdxSerialized.fullDescription} components={mdxComponents()}/>
                 : null}
 
               <div style={{ marginTop: "2rem" }}>
@@ -297,16 +188,13 @@ export default function ProjectPage({ projectData: project, mdxSerialized, sugge
             </div>
             <div>
               <H2>Omistajat</H2>
-              <AvatarRowExtended>
-                <AvatarRow members={project.members.map(m => ({ ...m, id: m._id, avatar: getMemberAvatarUrl(m._id) }))} />
-                <span>{project.members.map(member => member.name).join("; ")}</span>
-              </AvatarRowExtended>
+              <AvatarRow members={project.members.map(m => ({ _id: `ts:${m._id}`, name: m.name }))} withNames />
 
               {project.contributors.length > 0 ? <>
                 <H2>Kontribuuttorit
                   <Explanation>Projektin GitHub-repositorioihin koodia lisänneet, listattu GitHub-käyttäjät</Explanation>
                 </H2>
-                <AvatarRow members={project.contributors.map(c => ({ ...c, id: String(c.id) }))} />
+                <AvatarRow members={project.contributors.map(c => ({ _id: c.id, ...c }))} />
               </> : null}
 
               {project.links.length > 0 ? <>
@@ -317,7 +205,7 @@ export default function ProjectPage({ projectData: project, mdxSerialized, sugge
                       <Link href={link.url} >
                         <ProjectLinkTitleContainer>
                           {getProjectLinkIcon(link.type)}
-                          <h3>{getProjectLinkTitle(link.type, link.name)}</h3>
+                          <h3>{getProjectLinkTitle(link.type, link.name || "")}</h3>
                         </ProjectLinkTitleContainer>
                         <span>{getProjectLinkUrl(link.type, link.url)}</span>
                       </Link>
@@ -339,7 +227,7 @@ export default function ProjectPage({ projectData: project, mdxSerialized, sugge
           </Layout>
         </Content>
       </FadeBackground>
-      <Footer copyrightYear={copyrightYear} />
+      <Footer />
     </div>
   )
 }
@@ -353,6 +241,7 @@ export const getStaticProps: GetStaticProps<{
   suggestedProjectsData: ShallowProject[],
   copyrightYear: number
 }> = async (context) => {
+  if (context.params?.slug == undefined) throw "No slug"
   const { slug } = context.params
   if (typeof slug !== "string") throw new Error("Slug must be a string")
 
