@@ -1,7 +1,3 @@
-/**
- * https://github.com/Testaustime/testaustime-bot/blob/dcace5da793eeec3d61f36e6a768d993a87f61e2/src/lib/TimeUtil.js
- */
-
 export class TimeUtil {
     static Multipliers = {
         SECOND: 1,
@@ -19,6 +15,17 @@ export class TimeUtil {
         DAYS: "d ",
         WEEKS: "wks",
     };
+
+    static MonthNames = [
+        'tammikuuta', 'helmikuuta', 'maaliskuuta', 'huhtikuuta',
+        'toukokuuta', 'kesäkuuta', 'heinäkuuta', 'elokuuta',
+        'syyskuuta', 'lokakuuta', 'marraskuuta', 'joulukuuta'
+    ];
+
+    static WeekdayNames = [
+        'sunnuntai', 'maanantai', 'tiistai', 'keskiviikko',
+        'torstai', 'perjantai', 'lauantai'
+    ];
 
     /**
      * Format seconds into a readable string
@@ -61,5 +68,63 @@ export class TimeUtil {
                 )
                 .join(" ") || "0 secs"
         );
+    }
+
+    // Helper function to format the date as "d. mmmm." or "d. mmmm yyyy."
+    static formatFinnishDate = (date: Date): string => {
+        const day = date.getDate();
+        const month = this.MonthNames[date.getMonth()];
+        const year = date.getFullYear();
+        const currentYear = new Date().getFullYear();
+        
+        if (year === currentYear) {
+            return `${day}. ${month}`;
+        } else {
+            return `${day}. ${month} ${year}`;
+        }
+    };
+
+    /**
+     * Format date into a readable string in relation to the current date
+     */
+    
+    static formatDateInRelationToCurrent(date: Date): string {
+        const now = new Date();
+    
+        // Helper function to get the difference in calendar days
+        const differenceInCalendarDays = (date1: Date, date2: Date): number => {
+            const oneDay = 24 * 60 * 60 * 1000; // Milliseconds in a day
+            const diffDays = Math.round((date1.getTime() - date2.getTime()) / oneDay);
+            return diffDays;
+        };
+    
+        // Helper function to get the weekday name in Finnish
+        const getFinnishWeekdayName = (date: Date): string => {
+            return this.WeekdayNames[date.getDay()];
+        };
+    
+        const daysDifference = differenceInCalendarDays(now, date);
+    
+        if (daysDifference === 0) {
+            return 'tänään';
+        }
+        
+        if (daysDifference === 1) {
+            return 'eilen';
+        }
+    
+        if (daysDifference === 2) {
+            return 'toissa päivänä';
+        }
+    
+        const startOfWeek = new Date(now);
+        startOfWeek.setDate(now.getDate() - now.getDay() + 1); // Assuming week starts on Monday
+    
+        if (date >= startOfWeek && date <= now) {
+            const weekdayName = getFinnishWeekdayName(date);
+            return `${weekdayName}na`;
+        }
+    
+        return this.formatFinnishDate(date);
     }
 }
